nlp/requirements.txt
Flask==3.1.3
requests==2.32.5
python-dotenv==1.2.2
faiss-cpu==1.13.2
sentence-transformers==3.4.1
numpy==2.2.3
xgboost==2.1.3
scikit-learn==1.6.1
pandas==2.2.3
neo4j==5.28.1
networkx==3.4.2
